Slow ball speed to counter-act network latency
At multiplayer games the ball cannot collide with the opponent paddle.

With high latency, the ball would go through the opponent paddle,
and be reflected behind the opponent goal when opponent catch it.

As the opponent will take time to receive our paddle-ball collision,
the local ball will be ahead of opponents ball.

This commit will slow down (speed up) the local ball enough to align
it's position with opponent remote ball position.

So it can collide with opponent paddle at correct time.

Signed-off-by: Rafael Rondao <rafael.rondao@gmail.com>